Limit maximum childcare schedules to 4
