+++ src/main/resources/messages.properties
@@ -2046,6 +2046,8 @@ schedules-review.notice.no-providers-with-schedules.p1=You did not add any of yo
 schedules-review.notice.no-providers-with-schedules.p2=This application will be submitted for processing without any providers listed, unless you add them now.
 schedules-review.notice.providers-missing-schedules.p1=You did not add your providers, {0}, to the child care details.
 schedules-review.notice.providers-missing-schedules.p2=They will be removed from your application, unless you add them now.
+schedules-review.notice.more-than-4-children-needing-care.p1=Child care schedules can only be added for 4 children.
+schedules-review.notice.more-than-4-children-needing-care.p2=Your other children will be included on the application with limited care details.
 
 #schedules-same
 schedules-same.title=Same schedule
